// src/app/core/models/favorite-game.model.ts
import { Timestamp } from 'firebase/firestore';

export interface FavoriteGame {
  id: number;
  title: string;
  thumbnail: string;
  short_description: string;
  game_url: string;
  genre: string;
  platform: string;
  publisher: string;
  developer: string;
  release_date: string;
  freetogame_profile_url: string;
  addedAt: Timestamp;
}